+++ tests/testthat/test-workflow-plan.R
@@ -11,73 +11,95 @@ test_with_dir("empty plan", {
 })
 
 test_with_dir("plan set 1", {
-  x <- drake_plan(
-    a = c,
-    b = "c",
-    list = c(c = "d", d = "readRDS('e')"))
-  y <- data.frame(
-    target = letters[1:4],
-    command = c("c", "'c'",
-    "d", "readRDS('e')"),
-    stringsAsFactors = F)
-  expect_equal(x, y)
+  for (tidy_evaluation in c(TRUE, FALSE)){
+    x <- drake_plan(
+      a = c,
+      b = "c",
+      list = c(c = "d", d = "readRDS('e')"),
+      tidy_evaluation = tidy_evaluation
+    )
+    y <- data.frame(
+      target = letters[1:4],
+      command = c("c", "'c'",
+      "d", "readRDS('e')"),
+      stringsAsFactors = F)
+    expect_equal(x, y)
+  }
 })
 
 test_with_dir("plan set 2", {
-  x <- drake_plan(a = c,
-    b = "c",
-    list = c(c = "d", d = "readRDS('e')"),
-    strings_in_dots = "literals")
-  y <- data.frame(
-    target = letters[1:4],
-    command = c("c", "\"c\"",
-    "d", "readRDS('e')"), stringsAsFactors = F)
-  expect_equal(x, y)
+  for (tidy_evaluation in c(TRUE, FALSE)){
+    x <- drake_plan(a = c,
+                    b = "c",
+                    list = c(c = "d", d = "readRDS('e')"),
+                    strings_in_dots = "literals",
+                    tidy_evaluation = tidy_evaluation)
+    y <- data.frame(
+      target = letters[1:4],
+      command = c("c", "\"c\"",
+                  "d", "readRDS('e')"), stringsAsFactors = F)
+    expect_equal(x, y)
+  }
 })
 
 test_with_dir("plan set 3", {
+  for (tidy_evaluation in c(TRUE, FALSE)){
   x <- drake_plan(
     a = c,
     b = "c",
     list = c(c = "d", d = "readRDS('e')"),
-    strings_in_dots = "literals", file_targets = TRUE)
+    strings_in_dots = "literals", file_targets = TRUE,
+    tidy_evaluation = tidy_evaluation)
   y <- data.frame(
     target = drake::drake_quotes(letters[1:4], single = TRUE),
     command = c("c", "\"c\"", "d", "readRDS('e')"),
     stringsAsFactors = F)
   expect_equal(x, y)
+  }
 })
 
 test_with_dir("plan set 4", {
-  x <- drake_plan(
-    a = c,
-    b = "c",
-    list = c(c = "d", d = "readRDS('e')"),
-    strings_in_dots = "filenames", file_targets = TRUE)
-  y <- data.frame(
-    target = drake::drake_quotes(letters[1:4], single = TRUE),
-    command = c("c", "'c'", "d", "readRDS('e')"), stringsAsFactors = F)
-  expect_equal(x, y)
-  expect_warning(check_plan(x, verbose = FALSE))
+  for (tidy_evaluation in c(TRUE, FALSE)){
+    x <- drake_plan(
+      a = c,
+      b = "c",
+      list = c(c = "d", d = "readRDS('e')"),
+      strings_in_dots = "filenames", file_targets = TRUE,
+      tidy_evaluation = tidy_evaluation)
+    y <- data.frame(
+      target = drake::drake_quotes(letters[1:4], single = TRUE),
+      command = c("c", "'c'", "d", "readRDS('e')"), stringsAsFactors = F)
+    expect_equal(x, y)
+    expect_warning(check_plan(x, verbose = FALSE))
+  }
 })
 
 test_with_dir("drake_plan() trims outer whitespace in target names", {
-  x <- drake_plan(list = c(` a` = 1, `b \t\n` = 2))
-  y <- drake_plan(a = 1, b = 2)
-  expect_equal(x$target, y$target)
+  for (tidy_evaluation in c(TRUE, FALSE)){
+    x <- drake_plan(list = c(` a` = 1, `b \t\n` = 2),
+                    tidy_evaluation = tidy_evaluation)
+    y <- drake_plan(a = 1, b = 2, tidy_evaluation = tidy_evaluation)
+    expect_equal(x$target, y$target)
+  }
 })
 
 test_with_dir("make() and check_plan() trim outer whitespace in target names", {
   x <- data.frame(target = c("a\n", "  b", "c ", "\t  d   "),
-    command = 1)
+                  command = 1)
   expect_silent(make(x, verbose = FALSE, session_info = FALSE))
   expect_equal(sort(cached()), letters[1:4])
   stat <- c(a = "finished", b = "finished", c = "finished",
-    d = "finished")
+            d = "finished")
   expect_equal(progress(), stat)
 
-  expect_warning(make(x, verbose = FALSE, targets = c("a",
-    "nobody_home"), session_info = FALSE))
+  expect_warning(
+    make(
+      x,
+      verbose = FALSE,
+      targets = c("a", "nobody_home"),
+      session_info = FALSE
+    )
+  )
 
   x <- data.frame(target = c("a", " a"), command = 1)
   expect_error(check_plan(x, verbose = FALSE))